+++ ricaun.Nuke/Components/IHazGitHubActions.cs
@@ -1,0 +1,29 @@
+﻿using Nuke.Common.CI.GitHubActions;
+using Nuke.Common.IO;
+
+namespace ricaun.Nuke.Components
+{
+    /// <summary>
+    /// IHazGitHubActions
+    /// </summary>
+    public interface IHazGitHubActions
+    {
+        /// <summary>
+        /// GitHubActions
+        /// </summary>
+        GitHubActions GitHubActions => GitHubActions.Instance;
+
+        /// <summary>
+        /// GitHubSummary
+        /// </summary>
+        /// <param name="messages"></param>
+        void GitHubSummary(params object[] messages)
+        {
+            foreach (var message in messages)
+            {
+                Serilog.Log.Information($"GitHubSummaryFile: {message}");
+                GitHubActions?.StepSummaryFile.AppendAllText($"{message}\n");
+            }
+        }
+    }
+}